implementation dao and servise using crudrepository(not completted)

+++ src/main/java/app/repository/ClientRepository.java
@@ -1,9 +1,10 @@
 package app.repository;
 
 import app.libs.Client;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClientRepository extends CrudRepository<Client, Long> {
+public interface ClientRepository extends CrudRepository<Client, Integer> {
 }
