+++ src/main/java/app/repository/DriverRepository.java
@@ -1,10 +1,13 @@
 package app.repository;
 
 import app.libs.Driver;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.cdi.JpaRepositoryExtension;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DriverRepository extends CrudRepository<Driver, Long> {
+public interface DriverRepository extends CrudRepository<Driver, Integer> {
+
 
 }
