+++ src/main/java/app/repository/HistoryRepository.java
@@ -1,9 +1,11 @@
 package app.repository;
 
+import app.libs.Driver;
 import app.libs.History;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HistoryRepository extends CrudRepository<History , Long> {
+public interface HistoryRepository extends CrudRepository<History, Integer> {
 }
